Proper LaTeX parsing.
Among the advantages of using a LaTeX parser, `MString.fromLaTeX(...)`
is now a sensible thing to do. This dramatically improves searching over
imported bib files, since the represenation used for the search is
properly compiled from the LaTeX representation.

A side-effect is that `new` and `edit` now specifically work on the
LaTeX representation (as opposed to unicode before). We may introduce
additional parameters to these commands in the future if we see that
editing unicode is desirable.

Finally, having proper LaTeX parsing gave us a chance to fix import bugs
in the DBLP module (which uses a combination of JSON and BibTeX
extraction). Two examples of interest: 1) LNCS volume numbers are now
properly retrieved, and 2) fields that could not be retrieved from the
JSON result are now extracted from the bib entry (this is common for
some journal entries).