Catching up with DBLP changes in formatting.
This commit brings the DBLP module with the recent change in the
returned JSON format. It also introduces a "tolerance hack" in the JSON
parser, as many results are currently misformated by DBLP: it accepts
strings literals that are wrapped in (dictionary) curly braces. Finally,
it improves the error message when the JSON is definitely broken.